remove old packages view and refractor the extension list, disable the button when required
