+++ app/commands/keymaps.js
@@ -16,7 +16,6 @@ export default {
   'cmd+alt+shift+3': 'editor:split_pane_horizontal_3',
   'cmd+alt+4':       'editor:split_pane_vertical_4',
   'cmd+comma': 'global:show_settings',
-  'cmd+0': 'global:show_packages',
   'alt+b': 'global:show_branches',
 }
