+++ app/components/Modal/index.jsx
@@ -23,7 +23,6 @@ import {
   GitRebaseInput,
   GitCommitDiffView,
   GitCheckoutView,
-  PackageControlView
 } from './modals'
 
 var ModalContainer = (props) => {
@@ -116,9 +115,6 @@ class Modal extends Component {
 
         case 'Settings':
           return <SettingsView {...this.props} />
-
-        case 'Packages':
-          return <PackageControlView {...this.props} />
 
         default:
           return content
